## Changes

Improve `!bench server-stress` comment output. It should print a nicer markdown table now:

**Before:**


18k requests in 30.05s, 1.23 GB read


**After:**

|         |       Avg |     Stdev |     Max |
| :------ | --------: | --------: | ------: |
| Latency | 680.22 ms | 220.43 ms | 1398 ms |

|           |     Avg |   Stdev |     Min |   Total in 30s |
| :-------- | ------: | ------: | ------: | -------------: |
| Req/Sec   | 1449.54 |   73.74 |    1192 | 43.5k requests |
| Bytes/Sec |  103 MB | 5.24 MB | 84.8 MB |   3.09 GB read |


## Testing

<!-- How was this change tested? -->
<!-- DON'T DELETE THIS SECTION! If no tests added, explain why. -->
Tested by running `pnpm benchmark server-stress` locally.

## Docs

<!-- Could this affect a user’s behavior? We probably need to update docs! -->
<!-- If docs will be needed or you’re not sure, uncomment the next line: -->
<!-- /cc @withastro/maintainers-docs for feedback! -->

<!-- DON'T DELETE THIS SECTION! If no docs added, explain why.-->
<!-- https://github.com/withastro/docs -->
n/a. internal tooling.