@bluwy 2023-03-02
I've recently made the docs change at https://github.com/withastro/astro/pull/6396, a bit mixed up in that PR but I've got a couple reviews on it 😬